requirements.txt
certifi
charset-normalizer
idna
scikit-learn
pandas
Pillow
torch
pdfminer.six
regex
requests
sentencepiece
streamlit
tenacity
tiktoken
tqdm
transformers
urllib3
python-dotenv
dataclasses
python-docx
reportlab